Rag_agent/requirements.txt
langchain
langchain_community
langchain_huggingface
langchain_google_genai
langchain_groq
faiss-cpu
pypdf
huggingface_hub
flask
python-dotenv
sentence-transformers